add: 詳細画面（仮）への遷移追加
また、メインコンテンツの名称変更と分割
