+++ app/src/main/java/jp/co/yumemi/droidtraining/components/WeatherApp.kt
@@ -4,14 +4,10 @@ import WeatherFetchErrorDialog
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
@@ -21,24 +17,23 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
 import jp.co.yumemi.api.YumemiWeather
 import jp.co.yumemi.droidtraining.FakeWeatherMainViewModel
 import jp.co.yumemi.droidtraining.R
@@ -67,6 +62,8 @@ fun WeatherApp(
         },
     )
 
+    val navController = rememberNavController()
+
     Scaffold(
         topBar = {
             TopAppBar(
@@ -77,57 +74,32 @@ fun WeatherApp(
                 ),
             )
         },
-    ) {
-        WeatherAppContent(
-            modifier = Modifier.padding(it),
-            weatherInfoData = weatherInfoData,
-            onReloadClick = {
-                mainViewModel.reloadWeather()
-            },
-            enabled = !updating,
-        )
+    ) { innerPadding ->
+        NavHost(
+            navController = navController,
+            startDestination = Route.WeatherMain.name,
+            modifier = Modifier.padding(innerPadding)
+        ) {
+            composable(Route.WeatherMain.name) {
+                WeatherAppMainContent(
+                    weatherInfoData = weatherInfoData,
+                    onReloadClick = {
+                        mainViewModel.reloadWeather()
+                    },
+                    onNextClick = {
+                        navController.navigate(Route.WeatherDetail.name)
+                    },
+                    enabled = !updating,
+                )
+            }
+            composable(Route.WeatherDetail.name) {
+                WeatherAppDetailContent()
+            }
+        }
     }
 
     if (updating) {
         LoadingOverlay()
-    }
-}
-
-@Composable
-fun WeatherAppContent(
-    modifier: Modifier = Modifier,
-    weatherInfoData: WeatherInfoData,
-    enabled: Boolean = true,
-    onReloadClick: () -> Unit,
-) {
-    BoxWithConstraints(
-        modifier = modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center,
-    ) {
-        val screenWidth = with(LocalDensity.current) { constraints.maxWidth.toDp() }
-        val width = screenWidth / 2
-        Column(modifier = Modifier.width(width)) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .weight(1f)
-                    .padding(8.dp),
-                contentAlignment = Alignment.BottomCenter,
-            ) {
-                Text(text = weatherInfoData.place, fontSize = 20.sp)
-            }
-            WeatherInfo(weatherInfoData)
-
-            ActionButtons(
-                modifier = Modifier
-                    .padding(top = 80.dp)
-                    .weight(1f),
-                onReloadClick = {
-                    onReloadClick()
-                },
-                enabled = enabled,
-            )
-        }
     }
 }
 
@@ -254,15 +226,3 @@ fun PreviewAllWeatherApp(
     WeatherApp(mainViewModel = mainViewModel)
 }
 
-@Composable
-@Preview
-fun PreviewWeatherInfo() {
-    val weather = WeatherInfoData(WeatherType.SUNNY, 10, 20, "岐阜")
-    WeatherInfo(weatherInfoData = weather)
-}
-
-@Composable
-@Preview
-fun PreviewActionButtons() {
-    ActionButtons()
-}
