+++ app/src/main/java/jp/co/yumemi/droidtraining/components/WeatherAppDetailContent.kt
@@ -1,0 +1,9 @@
+package jp.co.yumemi.droidtraining.components
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun WeatherAppDetailContent() {
+    Text("Detail Content")
+}
